Don't start gitlab service until pids/ has been created
Sometimes the gitlab service will not start on the first provision. It works on the second attempt though.

Puma writes to /home/git/gitlab/tmp/pids/puma.pid. If pids/ doesn't exist, it fails silently. The order in which server.pp is run seems to cause the differing behaviour on fresh VMs.