Support for enabling backups
add configuration variables to support the enabling of a backup cron
job. This shell script will run gitlab's internal backup task:

  rake RAILS_ENV=production gitlab:backup:create

from cron. It is possible to also specify a script or a set of scripts
that run afterwards. This is useful to transfer the backup to a central
secure location. Gitlab backup job itself takes care of cleaning up old
files. The retention, in seconds, can be controlled via the
$gitlab_backup_keep_time variable, which defaults to 0 (forever).