Run migrations if gitlab is upgraded by `gitlab install`
At the moment only the `setup gitlab database` exec is run on
first run or until the `.gitlab_setup_done` file is deleted. But
upgrading gitlab normally requires running the db migrations.

This PR also adds some additional tasks to the `precompile assets`
exec that cleans the assets before precompiling as well as cleaning
the cache.

- Fixed spec test cases
- Fixed lint warnings