Charlock_holmes gem must require libicu-devel
The installation of charlock_holmes gem fails  on fresh installed systems
that does not have the libicu-devel.

This commit fix this problem making charlock_holmes gem require libicu-devel
package.

libicu package name is different on RHEL and Debian.

The method used for implementation is described bellow:
It was created another variable specific for the libicu package name,
just like it was done to git_package_name.
This method has the advantage of being less prone to errors and sideefect
free.