Managing postfix conflict
This wouldn't seem necessary because of the use of ensure_packages, but
unfortunately it's not a perfect function and only works if everyone
uses it. The thias/postfix module declares the Postfix package with
different parameters, and should probably take priority for the
package. This simple flag disables this module's (already really light)
management of the Postfix package if need-be.